chore(home): Move useGetTopFarmsByApr into its own hook
